+++ lino_welfare/modlib/integ/roles.py
@@ -1,5 +1,5 @@
 # -*- coding: UTF-8 -*-
-# Copyright 2015 Luc Saffre
+# Copyright 2015-2017 Luc Saffre
 # This file is part of Lino Welfare.
 #
 # Lino Welfare is free software: you can redistribute it and/or modify
@@ -21,9 +21,9 @@
 
 from lino_welfare.modlib.pcsw.roles import SocialAgent, SocialStaff
 from lino_xl.lib.cv.roles import CareerUser, CareerStaff
-from lino_welfare.modlib.courses.roles import CoursesUser, CoursesStaff
+from lino_welfare.modlib.xcourses.roles import CoursesUser, CoursesStaff
 from lino_welfare.modlib.newcomers.roles import NewcomersOperator
-from lino_cosi.lib.courses.roles import CoursesUser as CoursesUser2
+from lino_xl.lib.courses.roles import CoursesUser as CoursesUser2
 
 
 class IntegrationAgent(SocialAgent, CareerUser, CoursesUser, CoursesUser2):