+++ lino_welfare/modlib/xcourses/__init__.py
@@ -36,16 +36,16 @@ class Plugin(ad.Plugin):
     def setup_main_menu(self, site, profile, m):
         if True:  # profile.courses_level:
             m = m.add_menu(self.app_label, self.verbose_name)
-            m.add_action('courses.CourseProviders')
-            m.add_action('courses.CourseOffers')
-            m.add_action('courses.PendingCourseRequests')
+            m.add_action('xcourses.CourseProviders')
+            m.add_action('xcourses.CourseOffers')
+            m.add_action('xcourses.PendingCourseRequests')
 
     def setup_config_menu(self, site, profile, m):
         m = m.add_menu(self.app_label, self.verbose_name)
-        m.add_action('courses.CourseContents')
+        m.add_action('xcourses.CourseContents')
 
     def setup_explorer_menu(self, site, profile, m):
         m = m.add_menu(self.app_label, self.verbose_name)
-        m.add_action('courses.Courses')
-        m.add_action('courses.CourseRequests')
+        m.add_action('xcourses.Courses')
+        m.add_action('xcourses.CourseRequests')